Proxy API requests to seedbank-server (#66)
* Proxy API requests to seedbank-server

Add proxy support to both the Nginx container and the local development
environment. Modify the URLs of the API requests so they're sent to the same
server the JS code comes from, and it's then up to that server to forward them to
seedbank-server.

In local dev mode, requests are proxied to the URL specified in the
`REACT_APP_SEED_BANK_API` environment variable which is already needed to build
the code anyway. The value is pulled from the `.env` file if needed.

In the Nginx configuration, API requests are proxied to the URL specified in the
`SEEDBANK_SERVER_URL` environment variable, `http://seedbank-server:8080` by
default.

This eliminates the need for end users to edit their hosts files in order to use
the app, so `README.md` no longer needs instructions for doing that.

CU-mpymg6

* updates workflow and README

Co-authored-by: Carlos Thurber <carlos.thurber@admios-sa.com>