Trim columns from dump.sql INSERT statement (#154)
An upcoming change to seedbank-server will get rid of a few columns on the
accession table. Modify an INSERT statement in dump.sql so it only includes the
names of columns where it's inserting non-null data; none of those columns are
changing so this will make dump.sql continue to work after the schema change.