Regenerate types to reflect changes in the server API (#170)
* Regenerate types to reflect changes in the server API

UPDATING THE TYPE GENERATOR
---------------------------
When regenerating the types file, I got the following error:
  'AndNodePayload' is referenced directly or indirectly in its own type annotation.

This was caused by this following change, which introduced a
circular type reference:
  SearchNodePayload: (
    | components["schemas"]["AndNodePayload"]
    | components["schemas"]["FieldNodePayload"]
    | components["schemas"]["NotNodePayload"]
    | components["schemas"]["OrNodePayload"]
  ) & {
    operation: string;
  };

The following change to terraware server explains why the circular
type reference is valid and intended:
  https://github.com/terraware/terraware-server/commit/c2c9ebf99b34696224619abfde50a322add2d396

Therefore, I configured the openapi-typescript generator to
override the type of any node that contain the keyword
"TYPESCRIPT-OVERRIDE-TYPE-WITH-ANY".

CHANGES IN THE CLIENT CODE
--------------------------
I needed to update the client code in respose to the API changes.
Almost all the changes are related to species. I created a new
type: SpeciesType, which is used throughout the code rather than
the species types generated from the server API. This is
because we do not need access to all the fields that the server
returns. The new type also abstracts away the details of how the
API is called.

It looks like a lot of code was changed because I renamed 'specie'
to 'species' in many places. There are also some other minor
renames. Outside of the API calls, I didn't make any functional
changes to the client code.