SW-281 Use accession IDs in API calls; pass facility ID
We will need to support users who have access to more than one seed bank, which
means the API needs to be able to identify which seed bank's accessions to work
with.

Update the seed bank front end code to switch from accession numbers to accession
IDs as the unique identifiers in API calls. In addition, for API calls that don't
accept an accession ID, start passing in a facility ID.