Proxy requests to terraware-server
To emulate the behavior in the staging and production environments, where the
terraware-server API endpoints use the same hostname as the front-end code, add
proxy middleware to the dev server. With this setup, you can hit API and SSO
endpoints on port 3000.

This is of particular use when logging out: `http://localhost:3000/sso/logout`
will now redirect you back to the web app after logging out.