Skip E2E tests on main branch
The problem with Cypress tests hanging is making it pretty difficult to reliably
deploy code to staging after it's merged. Disable the tests on the main branch
until we can figure out a good fix/workaround.

Also remove a couple environment variables I'd set previously in an attempt to
fix the hangs; they didn't help.