SW-301 Attempt to work around Cypress hang
The E2E test suite always hangs in roughly the same place: updating or verifying
the seed count in the processing and drying test.

Waiting for the PUT request to finish and splitting the test step up into smaller
pieces appears to make it less likely for Cypress to hang.

But in case this doesn't fix the problem, also set a 1-hour timeout on the E2E
tests so jobs don't sit there for 6 hours chewing through our GitHub Actions
runtime quota.