Add logic to fetch organization data at the top level of the application
Create an async organization module that fetches all of a users's projects,
sites, facilities and layers. This module catches and parses all API errors,
so it will always return a promise that resolves. The return object will
specify what failed, if anything. If an error is specified, then partial
or no organization data will be returned.

Update the top level application rendering logic to show a simple error
message if the organization module fails to fetch projects or sites.

This change includes integration and unit tests for the organization module.

BREAKING CHANGES
----------------
Since this change does not remove the recoil selectors that manage plant
Layers, the Projects dropdown in the top nav bar is currently broken.
A follow up commit will fix this breaking change.