Restructure navigation bar + Ensure all paths navigate somewhere
Restructure the nav bar so that it's clear which views relate to
plants planted data vs seed bank data. Create placeholder
landing page (Homepage) that's not seed bank or plant specific.

Ensure that all paths navigate somewhere. Previously, the app
appeared broken if a user typed in a bogus path. Now,
  /home/bogus redirects to /home
  /seeds-summary/bogus redirects to /seeds-summary

Note that /accessions/bogus will not redirect, but it will show
the user an "accession not found" error, since "bogus" is
interpreted as an accession ID.

Also, fix some Material UI imports that were deprecated.