Integrate Plant Dashboard + Plant List with new APIS
Plants Dashboad and Plants List both take an Organization prop. They use
the organization data to fetch all the plant information that is needed
for each view.

When a plant is updated, the view reloads all of its plants data. This
behavior is identical to what was being done before. If the user is
navigates to a new view and then comes back to the dashboard, then we
refetch the plants data from the server. The same thing happens in the
Plants List. This is different from what was being done before. However,
making API calls when the user changes views should not be the performance
bottleneck here. The reload-on-edit behavior will most likely surface
performance issues first.

This diff removes all usages of Recoil, except for the snackbar (popup
that lets the user know an operation succeeded).

This diff also includes some changes to the networking modules.
  1. Update the getPlantSummary to return speciesId and plant counts,
     rather than just the total species count and plant count for a layer.
  2. Remove the speciesName from the Plant type and update the networking
     modules accordingly.

Things to address in a follow up diff:
  Displaying errors on the client side.
  Switching between projects, sites, facilities.