Create a networking-layer organization module that fetches all of a users's projects, sites, facilities and layers. This module catches and parses all API errors, so it will always return a promise that resolves. The return object will specify what failed, if anything. If an error occurred, then the organization data will be empty or incomplete.
This change includes integration and unit tests for the new module.
The tests are primarily focused on checking that the module handles
various error cases correctly.

Update the top level application rendering logic to show a simple error
message if the organization module fails to fetch projects or sites.
This UI is temporary. It will be updated once we have input from the
Design Team.

Include the jest tests in the list of CI tasks. Comment out the CI task
that generates typescript types because it was producing false-
positives.