Renamed some files and variables
Doing a second set of file renames in a follow up commit to keep the big commit clean.

Updated names like filters2 to filters. I intentionally used a differnt name in the main
refactor diff so that the code diffing is more obvious. Filters2 has a pretty different
usage compared to filters and I didn't want that to be confusing.

Also, deleted some unused code in snackbar.