Refactor Plant Dashboard and Plant List (#78)
This diff tackles 4 main things.

1. Creates a number of networking modules that make API calls, parse the server response
into new client-side types, and return abstract errors to the caller. The functions in these
networking modules are async but they ALWAYS return a promise that resolves. That
allows the client code to safely use await getPlants() without needing to wrap the call in
a try/catch block. The caller must examine the response to check that the request
succeeded.

2. Introduces unit testing into the code base. All the new networking modules include Jest
tests, which are now included in the CI flow.

3. Reorganizes the site navigation and ensures that all urls will resolve. I combined this
change with the refactor because it had some data flow implications.

4. Utilizes the new networking modules to redefine how data flows through the application.
Organization data is fetched at the top level of the app and passed down to the
PlantDashboard and PlantList modules (formerly called Dashboard and AllPlants).
The plant dashboard and plant list views are now responsible for maintaining their own
plants and species state (rather than sharing state using recoil).

Additionally, this diff includes a number of smaller changes:

-- Makes some small tweaks to the existing end to end tests. Removes a couple of tests
that are no longer relevant.
-- Recreates the yarn.lock because Typescript was complaining about that we're using an
incorrect version. That incorrect version was most likely being specified in a dependency.
Changing the typescript version caused a few type errors in existing code so I added
ts-ignore to those for now.
-- Fixes some deprecated Material UI imports.
-- Comments out the CI tasks that attempts to generate the API types file. That tasks was
failing silently.

This is a very large change that was originally broken up into 8 relatively clean commits
to make them easier to review. Those commits will be squashed when this diff is merged
into main, but the history can always be accessed in the original PR.