Run Prettier formatter on (almost) the whole repository (#82)
PR #79 formalized the use of Prettier in the codebase. However, that was only
for the /src directory. Expand the use of Prettier to (almost) the entire repository.

Update .prettierignore to exclude all directories we don't want formatted
Update yarn format command to run Prettier on the entire repository
Update CI check to run on the entire repository
Run updated yarn format command