User can navigate away from plants list view and filters will be preserved (#80)
During a refactor, #78 removed the ability to preserve filters on the plants list
page after a user navigates to a new view. This change re-enables that behavior.
Thanks @constanzauanini for pointing out that this behavior is a product
requirement.

I included two very basic unit tests that check the PlantFilterBar renders
correctly. In a future diff, these tests could be replaced with a more
comprehensive integration test.

I also re-added the end to end test that checked this behavior before #78.