Remove (seedbank) species Recoil selector
Since we used to have a separate codebase for the plant locator and
seedbank functionality, there were two different networking modules
being used to fetch species information. This diff migrates all of
the seedbank code over to the most recently updated species
networking module, which has error handling.

As a result, we're able to remove the unnecessary species recoil
selector that was being used by the seedbank flows. We're also able
to delete the less fault tolerant species networking module.