TRACK-514 Use alternate names for search fields
As part of reworking the search API, we introduced alternate names for four
accession search fields. Update the front end to use them instead of the old
names so we can remove the old names on the back end. The renames:

* `family` -> `familyName`
* `primaryCollector` -> `primaryCollectorName`
* `species` -> `speciesName`
* `storageLocation` -> `storageLocationName`

This only affects searches, not other API endpoints where those pieces of
information are passed around.

On the back end, we've removed the list of search fields from the OpenAPI schema.
That'll affect the generated front end code as well. However, to keep the scope
of this change smaller, I instead hand-edited `generated-types.ts` to update the
names.