Remove search selectors (#94)
Remove the Recoil accessions search selectors and replace them with React slices
of state. In some cases, that state could be local to the components using it. However,
I did need to lift some state up into App.tsx so that it can be cached when the user
switches between views.

Putting these slices of state in the top level of the application makes it more obvious
how they are being used. However, there is some overhead of passing that state as
props between components.

This change also renames many of the search related function & variable names. It
improves the search code documentation generally. This is at least as important as
removing the recoil usage.

Finally, this diff also consolidates some auto-generated types files into the networking
modules where they are being used. Specifically

  - src/api/types/locations.ts --> src/api/seeds/locations.ts
  - src/api/types/report.ts    --> src/api/seeds/reports.ts
  - src/api/types/search.ts    --> src/api/seeds/search.ts
  
This in the final major change that's included in the task of removing Recoil usage and
enabling our API modules to handle errors gracefully. The other PRs included in this task
are #78, #80, #84, #86, #87, and #89.