TRACK-421 Expect old families in Cypress tests
An upcoming terraware-server change is going to make the family name a text field
on the accession, rather than a reeference to a separate list of families. That
will have the side effect of changing the behavior of the summary endpoint such
that, given the test data set we use for the Cypress tests, it will report that
there was already a family as of the previous week.

In anticipation of that change, update the data set on the existing data model so
the example family was created in the past, and adjust the tests to expect to
see a "since last week" delta.