TRACK-528 Reduce dependence on server database schema
Replace the hand-written database setup script with a dump of the server database
after running the previous version of the script. The dump includes metadata about
which schema migrations have been run. When terraware-server starts up, if there
are additional schema changes, the appropriate migrations will run.

This should reduce the number of times the front-end test suite is inadvertently
broken by updates to the back end.

As part of this, `scripts/resetdb.sh` is reorganized a bit so it restores the
database dump before starting terraware-server. That's needed because
terraware-server only runs database migrations when it starts up.