Fix ID numbering collisions in database dump (#125)
The database dump inserts some initial data with a fixed set of IDs, but it didn't
update the sequences that are used to allocate IDs for newly-inserted data. This
caused failures when creating certain kinds of objects, e.g., the first attempt to
create an organization would fail because it would try to use organization ID 1
which had already been inserted.

Update the starting values of the ID sequences for tables that are preloaded with
starting rows in the database dump.