Show create project/site message when no sites (#160)
On plants list and plants dashboard we show a message to create a project and
site when there are no projects or sites created. Previously it was showed only
when no projects were created, and now we are adding the no sites condition
also. We are also hiding the projects dropdowns when no projects or sites are
created.