Improve condition to clear selected rows (#175)
https://terraformation.atlassian.net/browse/TRACK-755

The condition to clear the selected rows was anytime the rows changed,
but we were checking that at least one row exist for reseting the selected
rows. This check was to avoid an infinite loop that happens when there are
no rows. The problem with this condition was that when the rows use to
have length bigger than 0 and then rows became no length, the selected
rows were not refreshing, so a new condition is added to fix this issue.