Update UI to reflect that Admins and Owners are part of all Projects (#185)
TRACK-764
- When creating, editing, or viewing an Admin or Owner's profile, don't show a
list of projects. Instead, show message saying that "Admins and owners
automatically have access to all Projects".
- When creating, editing, or viewing a project, don't show admins or owners
in the list of people that can be added to the project. Update the UI to reflect
that you can only add contributors to a project because admins and owners
automatically have access to all projects.

TRACK-800
- Prevent a user from changing the organization role of any Owner. This is
necessary because we currently don't have a way to assign a new owner.