Update record with accession only if updating (#223)
Because we refresh the notifications each minute, when creating a new accession, the new accession page is refreshed each minute, and a "new accession" was passed each minute, which cases the user to lose their progress when filling out the new accession form. We need to avoid updating the accession with the new accession to fix this issue.