SW-1064 Pass health check requests to backend
We need the load balancer to be able to detect when terraware-server isn't
working. Add a passthrough route for the Spring Boot health check endpoint
to the Nginx configuration. The load balancer will start using this, which will
verify both the backend's health and the health of Nginx (since the request
won't get forwarded if Nginx isn't running).