SW-693 Authenticate with AWS using OpenID Connect (#393)
Update the GitHub Actions workflow to stop using an access key to authenticate
with AWS during deployment. Instead, GitHub will generate a signed OpenID
Connect token and use it to request temporary access to an AWS role that only
allows access to the specific operations we need for deployment.

This will eliminate the need to rotate our access key periodically.