Remove unnecessary steps in CI builds (#395)
We don't run Cypress tests when a release is tagged, so there's no need to
download and start terraware-server either. This will speed up release builds by
30-40 seconds.

And in non-release builds, there's no point in trying to prepopulate the database
with plant location data since the endpoints in question don't exist any more.