Update database dump with v2 accession migration
https://github.com/terraware/terraware-server/commit/fe3e5b1276 introduced a
migration step that only existed for long enough to run it in production, and
was removed afterwards because it relied on obsolete code we needed to remove.

Apply this migration to the test database used by the Cypress test suite so
that the suite can be run with more recent server versions.

Also add some documentation to help server developers update the database dump
in the future.