SW-2338, SW-2339 Withdrawal/Reassignment Table (#923)
Table calculations:
On the withdrawal tab: only show quantities for the original withdrawal,
and do not take into account reassignments. This is because the user
wants to know what was originally withdrawn from the nursery and expects
to go to the reassignment tab to find more information about whether
plants ended up in different plots than expected.

On the reassignment tab: show a row on the table for each plot that a
species is delivered to. The Original Qty column is essentially identical to
the quantities on the withdrawal tab and shows how many plants are
originally intended for the plot. The Final Qty tab shows how many ended
up in the plot at the end of the day, accounting for changes due to
reassignments.