SW-2714 Simplify "has withdrawals?" search request
At page load, we query the backend to find out if there are any nursery
withdrawals; if not, we don't show the "Withdrawal Log" item in the left nav.
Previously, this was searching for all the withdrawal details, but really we
only need to know if a withdrawal exists. Change it to just search for the ID.