Move English strings to a separate source file
In preparation for dynamically loading localized strings, move the English strings
to a separate source file and tighten up the type declarations of the strings
table. For now, the new file is loaded statically, so there's no change in
behavior here; it will still be included in the Webpack bundle as before.