SW-2721 Look up column labels at render time (#1042)
Update the table views so they look up the display names of the columns at the
time the render function is called rather than at the time the module is imported.

For most of the tables, this is a simple matter of moving the column list from the
top level of the module into the body of the render function.

The list of columns on the accessions table, though, is referenced in multiple
places and is also pretty large. Change it from a constant to a function and call
the function in the places that used to reference the constant.

No user-visible behavior change here; the English strings table is still imported
statically as before.