SW-2620 Filter Pill List Component
Create a PillList component that renders a list of pills. Also make the
Pill component more generic. Also, if the pill is not rendered with an
onRemovePill handler it will not have an 'X' icon to delete it.

Refactor existing components that display pills (table filters) to use
the new PillList component.