SW-2620 Filter Pill List Component (#1060)
Create a PillList component that renders a list of pills. Also make the
Pill component more generic. If the pill is not rendered with an
onRemove handler it will not have an 'X' icon to delete it.

Refactor existing components that display pills (table filters) to use
the new PillList component.