SW-2733 Pass localization parameters to components (#1068)
Some of the components in web-components can display localizable strings.
Update the application to pass the necessary strings and render functions
to the components so they will render in the user's locale.

The general approach here is to put the localization parameters in wrapper
functions that pass through to the actual components. Most of the changes
are thus replacing imports of the underlying components with imports of
the wrappers.