Fix flapping when non-English locale is selected
Commit fe1261459 introduced a bug: if the user's profile had a non-English locale,
the UI would flap back and forth between English and the selected locale.

This was because LocalizationProvider was treating any update of the locale as a
trigger to save the new value to the user's profile. So when the user's profile
was loaded and the locale was set based on that, it triggered a write to the
user's profile, which then triggered another locale update, and so on.

Fix it by only updating the user's profile when they actually change the locale
selector.