SW-2778 Reset searches when locale changes (#1127)
The People page and the Accessions page include localizable search results.
When the user changes locales, we need to show the results in the new locale.

In the case of the Accessions page, we also need to reset the search criteria
because they might include localized values in the old locale, which the
search API won't recognize in the new locale. The People page doesn't allow
filtering on the localizable column (role) so doesn't need this logic.