SW-2806 Localize notification timestamps
Currently, the timestamps on notifications are rendered using MomentJS with a
fixed format string. Replace that with a component that renders a timestamp
using the browser's built-in `Intl.DateTimeFormat` class. Since it's not possible
to add custom locales to `Intl`, the component has some special-case code for
the gibberish locale that can be used to test that dates are in fact localized.

The default US English date format is similar to the one we were using before,
but not identical.

- Old: `January 21st 2023 at 12:14:54 am`
- New: `January 21, 2023 at 12:14:54 AM`