SW-2869 Add component for text with embedded link
To make it easier to translate strings with embedded links, add a new component
that replaces delimiter characters with the HTML link tags. This will allow
translators to use the grammatically correct text for the link based on its
surrounding context.

This initial implementation is pretty limited; it only supports a single link
per string.

We no longer need the `html-react-parser` library, which was only being used
to construct an embedded link.