Revert index.html change that removed lang attribute (#1151)
The W3C recommendation is to always include a `lang` attribute on `<html>`
tags.  We can't set it to the user's locale in our static `index.html` since
it's not known until after we fetch the user's profile information from the
server, so hardwire it to English on initial page load. The default will be
overwritten by the application code once it loads the user's information.