SW-2906 Improve localization of accession details (#1155)
The accession details page had a hardwired English menu item (Delete) and was
only rendering the age at initial component render time, so it wasn't updating
when the user switched locales.

Even in English, the accession age wasn't quite ideal; it showed a value of
"1 Months" instead of "1 Month". Add support for singular month values and move
the concatenation of the number of months and the word "Months" into the strings
table to support languages with different word order.