SW-2618 Build Generic Filter Edit Dialog Component (#1185)
Create a FilterGroup component that takes a list of filter field with
specified filter types (e.g. multiple_selection, number_range, etc).
Also take a FieldValuesPayload containing the options for each field, if
applicable. The FilterGroup may be rendered inside whichever container is
desired, such as a dialog or a popover. TODO: one item that needs to be
done still is to render the multiple_selection filter type as a Filter-
MultiSelect. Change the single_selection component to a dropdown.
These will be done in a followup.

The FilterGroup is implemented in the Accessions table page, to show it
in action. We remove the chip-based filters and also confirmed that each
of the types of filter works correctly. TODO: add pills showing active filters.