SW-3272 Generate string tables from CSV files
The translation service we're going to use doesn't support TypeScript source
files. Move the English strings to a CSV file that has an additional column
(currently empty for all our strings) where we can add comments to give additional
context to translators.

We will send the English CSV file out for translation, and get back CSV files in
other languages; these will then be checked into the repo.

There's now a build step that converts the CSV files to JavaScript so they can be
imported from the rest of the code as usual. This is done as an explicit one-off
step in the CI workflow; in dev environments, the CSV files are monitored for
changes and processed on the fly so that edits to strings will be hot-reloaded
the same way they were reloaded when the strings table was a TypeScript source
file.

This change doesn't introduce any actual translation workflow steps, and for now,
the only locale with a CSV file is US English. But this will enable us to start
adding comments to our strings right away.