SW-3344 Prevent Browser Back When Editing Report (#1331)
When editing a report, allowing the back button to fire a navigation can
result in work being lost. Here, we disable that action when in edit
mode, and display a warning to the user to use the form controls
instead. We also refactor the type of the switch pages function in
ReportEdit.