SW-3272 Include all English keys in JS strings (#1351)
If we have CSV files for multiple locales, we generate a JavaScript source file
with the strings for each locale. Previously, this was just an alternate rendering
locale's CSV file.

Unfortunately, that simple approach means that if you add or remove an English
string, the JS strings objects for the other locales won't have the same keys as
the English one. At best, this will cause a TypeScript error because the type
signatures don't match; at worst, it'll be a runtime error.

Instead, generate each locale's JS file using both the English and localized CSV
files. The list of keys will always be based on the English file. The values will
be the localized strings if available, or the original English strings if not,
such that the app effectively falls back to English for strings that aren't
translated yet.

Print a warning for each untranslated string; that'll let us easily check the CI
build logs before a release to see if we're missing any translations.