Planting Site Map Data (#1438)
We change the way that map data flows into the PlantingSiteMap. Instead of
passing in the planting site, we pass in a more generic MapData object, which
contains data for the various map objects (site, zone, subzone, permanentPlot,
temporaryPlot). This will help make the component more reusable where, for
instance, one does not have a planting site object, but observation results
from which the map data object may be created. Behavior for existing maps, such
as the ones in the Planting Site or Plants Dashboard pages, is unchanged.