SW-3348 Configure Spring Security OAuth2 on server
Update the local terraware-server configuration to reflect the fact that the
server is switching to a different authentication library. Developers will
have to modify their local `.env` files to move some of the configuration to
different environment variables; `.env.sample` is updated to show what the
configuration needs to look like.