SW-3818 Add new growth forms
The forestry team requested some new growth forms; add them.

Start sorting the list of growth forms in drop-down menus and the
Learn More popup to make it easier for users to find the growth form
they're looking for, since the list is now too long to take in at a
glance all at once.